+++ packages/core/src/modules/user/user-sync-service.ts
@@ -3,22 +3,50 @@ import type { User, UserWrite } from "@dotkomonline/types"
 import { addDays } from "date-fns"
 import { Auth0UserNotFoundError } from "../external/auth0-errors"
 import type { Auth0Repository } from "../external/auth0-repository"
-import type { UserRepository } from "./user-repository"
+import type { UserService } from "./user-service"
 
 export interface UserSyncService {
   update(payload: UserWrite): Promise<User>
   handleUserSync(sub: string, now: Date): Promise<User>
+  handlePopulateUserWithFakeData(auth0Id: string, email?: string | null): Promise<void>
+}
+
+// Until we have gather this data from the user, this fake data is used as the initial data for new users
+const FAKE_USER_EXTRA_SIGNUP_DATA: Omit<UserWrite, "email" | "id" | "auth0Id"> = {
+  givenName: "firstName",
+  familyName: "lastName",
+  middleName: "middleName",
+  name: "firstName middleName lastName",
+  allergies: ["allergy1", "allergy2"],
+  picture: "https://example.com/image.jpg",
+  studyYear: -1,
+  lastSyncedAt: new Date(),
+  phone: "12345678",
+  gender: "male",
 }
 
 export class Auth0UserSyncService implements UserSyncService {
   private readonly logger: Logger = getLogger(Auth0UserSyncService.name)
   constructor(
-    private readonly userRepository: UserRepository,
+    private readonly userService: UserService,
     private readonly auth0Repository: Auth0Repository
   ) {}
 
+  async handlePopulateUserWithFakeData(auth0Id: string, email?: string | null) {
+    if (!email) {
+      throw new Error("Did not get email in jwt")
+    }
+    const user = await this.create({
+      ...FAKE_USER_EXTRA_SIGNUP_DATA,
+      email: email,
+      auth0Id: auth0Id,
+    })
+
+    await this.update(user)
+  }
+
   async create(data: UserWrite) {
-    return this.userRepository.create(data)
+    return this.userService.create(data)
   }
 
   async update(data: UserWrite) {
@@ -30,20 +58,20 @@ export class Auth0UserSyncService implements UserSyncService {
   async syncUser(userAuth0: User) {
     this.logger.log("info", "Synchronizing user with Auth0 id", { userId: userAuth0.auth0Id })
 
-    const updatedUser: UserWrite = {
+    const updatedUser: User = {
       ...userAuth0,
       lastSyncedAt: new Date(),
     }
 
-    const userDb = await this.userRepository.getById(userAuth0.id)
+    const userDb = await this.userService.getById(userAuth0.id)
 
     if (userDb === null) {
       this.logger.log("info", "User does not exist in local db, creating user for user ", userAuth0.name)
-      return this.userRepository.create(updatedUser)
+      return this.userService.create(updatedUser)
     }
 
     this.logger.log("info", "Updating user in local db for user ", userAuth0.name)
-    return this.userRepository.update(userAuth0.id, updatedUser)
+    return this.userService.update(updatedUser)
   }
 
   /**
@@ -52,7 +80,7 @@ export class Auth0UserSyncService implements UserSyncService {
    * @returns User
    */
   async handleUserSync(auth0UserId: string, now: Date) {
-    const userDB = await this.userRepository.getByAuth0Id(auth0UserId)
+    const userDB = await this.userService.getByAuth0Id(auth0UserId)
 
     const oneDayAgo = addDays(now, -1)
     const userDoesNotNeedSync = userDB?.lastSyncedAt && oneDayAgo < userDB.lastSyncedAt
