+++ packages/gateway-trpc/src/modules/user/user-router.ts
@@ -14,7 +14,7 @@ export const userRouter = t.router({
         data: UserWriteSchema,
       })
     )
-    .mutation(async ({ input: changes, ctx }) => ctx.userService.updateUser(changes.data)),
+    .mutation(async ({ input: changes, ctx }) => ctx.syncedUserService.update(changes.data)),
   getPrivacyPermissionssByUserId: protectedProcedure
     .input(z.string())
     .query(async ({ input, ctx }) => ctx.userService.getPrivacyPermissionsByUserId(input)),
